+++ src/Core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Fixed
+
+- Fix invalid signature in SES client because of wrong Scoped Service.
+
 ## 1.2.0
 
 ### Added